Moved extract_age_numeric function to utils so other transforms can use it
