+++ tests/transform/test_disease_correlation.py
@@ -6,7 +6,6 @@ from agoradatatools.etl.transform.disease_correlation import (
     transform_disease_correlation,
     create_lookup,
     extract_module_name,
-    extract_age_numeric,
     process_group,
 )
 
@@ -623,33 +622,3 @@ class TestProcessGroup:
 
         # Should take first element from the list
         assert result["matched_control"] == "C57BL6J"
-
-
-class TestExtractAgeNumeric:
-    """
-    Test class for validating the extract_age_numeric utility function.
-    This function extracts the numeric value from an age string.
-    """
-
-    @pytest.mark.parametrize(
-        "input_age,expected",
-        [
-            ("4 months", 4),  # Age with months
-            ("8 months", 8),  # Age with months
-            ("12 months", 12),  # Age with months
-            ("6 weeks", 6),  # Age with weeks
-            ("100 days", 100),  # Age with days
-            ("", 0),  # Empty string
-            ("no number here", 0),  # String without number
-        ],
-    )
-    def test_extract_age_numeric(self, input_age, expected):
-        """
-        Test that extract_age_numeric correctly extracts the numeric value
-        from age strings with various formats.
-
-        Args:
-            input_age: Input string that may contain age and unit
-            expected: Expected numeric age value
-        """
-        assert extract_age_numeric(input_age) == expected
